+++ runtime/src/commonMain/kotlin/dev/zacsweers/metro/internal/MapFactory.kt
@@ -21,7 +21,7 @@ import dev.zacsweers.metro.Provider
  * A [Factory] implementation used to implement [Map] bindings. This factory returns a `Map<K, V>`
  * when calling [invoke] (as specified by [Factory]).
  */
-public class MapFactory<K : Any, V : Any> private constructor(map: Map<K, Provider<V>>) :
+public class MapFactory<K : Any, V> private constructor(map: Map<K, Provider<V>>) :
   AbstractMapFactory<K, V, V>(map) {
 
   /**