+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changelog
 **Unreleased**
 --------------
 
+- **New**: Nullable bindings are now allowed! See the [nullability docs](https://zacsweers.github.io/metro/bindings#nullability) for more info.
+- **Enhancement**: Add diagnostics for multibindings with star projections.
+- **Enhancement**: Add diagnostic for map multibindings with nullable keys.
+
 0.1.3
 -----